+++ lib/ext/string.dart
@@ -3,3 +3,7 @@ extension StringExtension on String {
     return "${this[0].toUpperCase()}${substring(1).toLowerCase()}";
   }
 }
+
+extension StringExtensions on String {
+  bool containsIgnoreCase(String secondString) => toLowerCase().contains(secondString.toLowerCase());
+}